remove _core_C/_core_ext.py and move Scalarype to scalar_type.py
