+++ vllm/_custom_ops.py
@@ -6,9 +6,9 @@ import torch
 import torch.library
 
 import vllm.envs as envs
-from vllm._core_ext import ScalarType
 from vllm.logger import init_logger
 from vllm.platforms import current_platform
+from vllm.scalar_type import ScalarType
 
 logger = init_logger(__name__)
